Editor inherits from additional classes: listener and timer
The listener class will allow listening to parameters. When changed, the editor will be able to update the UI and change its EQ graph (for now the graph does nothing even when we tweak knobs). Timer is needed because we can't update the GUI when a parameter is changed, it's too slow. So timer will check a flag and update UI in its own thread